requirements.txt
python-dotenv==0.13.0
spotipy==2.12.0
Flask==1.1.2
requests~=2.23.0
